# Streamline HTML generation and consolidate templates