--- a/build-scripts/webpack.cjs
+++ b/build-scripts/webpack.cjs
@@ -84,6 +84,7 @@ const createWebpackConfig = ({
       ],
       moduleIds: isProdBuild && !isStatsBuild ? "deterministic" : "named",
       chunkIds: isProdBuild && !isStatsBuild ? "deterministic" : "named",
+      runtimeChunk: "single",
     },
     plugins: [
       !isStatsBuild && new WebpackBar({ fancy: !isProdBuild }),