… cast/src/html/launcher.html.template
… gallery/src/html/index.html.template
… hassio/src/entrypoint.js.template
… src/html/_header.html.template
… src/html/_script_load_es5.html.template
… src/html/authorize.html.template
… src/html/index.html.template
… src/html/onboarding.html.template
… test/webpack.config.js