<!--
  You are amazing! Thanks for contributing to our project!
  Please, DO NOT DELETE ANY TEXT from this template! (unless instructed).
-->

## Proposed change

<!--
  Describe the big picture of your changes here to communicate to the
  maintainers why we should accept this pull request. If it fixes a bug
  or resolves a feature request, be sure to link to that issue or discussion
  in the additional information section.
-->
_**Note:** The runtime chunk changes have been removed due to issues with Webpack.  See discussion below._

The main goal here is to separate Webpack's runtime code into its own chunk per [Webpack's caching guide](https://webpack.js.org/guides/caching/#extracting-boilerplate).  Currently that code is bundled in with each entrypoint's initial chunk.  Since it includes an internal manifest that changes each version, the large entry chunks can never be long-term cached.  This fixes that.

In addition, since the index page of the main app includes 2 entrypoints, the runtime is also being included and instantiated twice there.  This fixes that too with a single runtime shared amongst all entrypoints.

Now the fun part... the `entry-html` tasks and templates already have quite a bit of repetitive code, so rather than make that much worse, I completely overhauled those tasks and tried to consolidate some template HTML.  Here's the rough summary:

- Two main functions (dev and prod) now control the tasks.  Each project just gives it the proper paths and a map of pages to entrypoint(s).  This also means that several that were not being minified now are.
- Nested templates are now passed the same data and have relative paths so they can work anywhere
- Consolidated the ES5 loading into a child template used by many pages.  The latest loads were too much of a hodgepodge so maybe next time.
- Got rid of supervisor function writing code and just made it a template

There's certainly more that can be streamlined here, but this is a good start I think.

## Type of change

<!--
  What type of change does your PR introduce to the Home Assistant frontend?
  NOTE: Please, check only 1! box!
  If your PR requires multiple boxes to be checked, you'll most likely need to
  split it into multiple PRs. This makes things easier and faster to code review.
-->

- [ ] Dependency upgrade
- [ ] Bugfix (non-breaking change which fixes an issue)
- [x] New feature (thank you!)
- [ ] Breaking change (fix/feature causing existing functionality to break)
- [x] Code quality improvements to existing code or addition of tests

## Example configuration

<!--
  Supplying a configuration snippet, makes it easier for a maintainer to test
  your PR.
-->

```yaml

```

## Additional information

<!--
  Details are important, and help maintainers processing your PR.
  Please be sure to fill out additional details, if applicable.
-->

- This PR fixes or closes issue: fixes #
- This PR is related to issue or discussion:
- Link to documentation pull request:

## Checklist

<!--
  Put an `x` in the boxes that apply. You can also fill these out after
  creating the PR. If you're unsure about any of them, don't hesitate to ask.
  We're here to help! This is simply a reminder of what we are going to look
  for before merging your code.
-->

- [x] The code change is tested and works locally.
- [x] There is no commented out code in this PR.
- [ ] Tests have been added to verify that the new code works.

If user exposed functionality or configuration variables are added/changed:

- [ ] Documentation added/updated for [www.home-assistant.io][docs-repository]

<!--
  Thank you for contributing <3
-->

[docs-repository]: https://github.com/home-assistant/home-assistant.io