@steverep 2023-04-10
Had to disable the runtime chunk manually for the mocha tests to avoid an error.  It's caused by instant-mocha overriding the output filename to `"main.js"`.  The runtime chunk is irrelevant for unit tests anyway.

Ref: https://github.com/privatenumber/instant-mocha/issues/47