@balloob 2023-04-11
The `core.ts` entrypoint needs to load as fast as possible and not rely on other files in the process. It's purpose is to start the loading of data while the browser is busy processing the JavaScript that builds up the frontend. Therefore, it should not have its webpack runtime extracted.

Another thing to note is that Home Assistant is served in two ways: locally without HTTPS and remotely with HTTPS. When it's local, the connection speed is fast. Caching might not be super important. When it's remote, connection speeds are often slow, but caching is managed by the service worker which pre-downloads all the files locally.