@steverep 2023-04-11
> The `core.ts` entrypoint needs to load as fast as possible and not rely on other files in the process. It's purpose is to start the loading of data while the browser is busy processing the JavaScript that builds up the frontend. Therefore, it should not have its webpack runtime extracted.

For modern builds, they are both loaded async, so core is not waiting for anything, right?  The only difference is the extra file.  Plus, core is dependent on the runtime to load fully anyway because it controls the loading of any split chunks.

> Another thing to note is that Home Assistant is served in two ways: locally without HTTPS and remotely with HTTPS. When it's local, the connection speed is fast. Caching might not be super important. When it's remote, connection speeds are often slow, but caching is managed by the service worker which pre-downloads all the files locally and updates in the background.

Maybe I'm missing your point here.  The goal is to allow the service worker to cache the large entry/initial chunks longer.  Currently they get busted anytime a new frontend is released even though most of the time the actual changes are much deeper in the tree.