@bramkragten 2023-04-11
Core should (ideally) not have any split chunks. It should all be in the entrypoint