@steverep 2023-04-11
> Core should (ideally) not have any split chunks. It should all be in the entrypoint

Just checked my production build and that looks to be correct.  It stands alone. 👍🏻  However, it still depends on the runtime code because of all the `__webpack*` that replaces the imports and exports.

Still, I'm failing to see where any measurable slow down would occur as a result of this change.  There might even be a slight speed up under normal circumstances, and certainly should be faster after an update or when coming from the login or onboarding pages.