@balloob 2023-04-11
It loads and starts parsing `core.ts`, then right away realizes it needs to load another chunk and needs to initiate another request and wait for that to resolve.

Maybe the caching in the server worker has changed, but I was under the impression that it would throw away all the cache of the old service worker version and download a fresh copy.