@steverep 2023-04-11
> It loads and starts parsing `core.ts`, then right away realizes it needs to load another chunk and needs to initiate another request and wait for that to resolve.

I think maybe you're misunderstanding?  Core doesn't load the runtime.  The runtime chunk is an initial chunk just like the other entrypoints.  It gets preloaded and dynamically imported asynchronously alongside the others from the HTML (hence the bulky changes).  In fact, you could await the core and app loads before importing the runtime, but I don't see much point there.

So any performance loss would come from just the overhead of dealing with 1 extra file and, absent cache, 1 extra HTTP request.  But for the index page, that gets offset by the facts that:
- It's preloaded while the page is parsed
- About 270 lines of repeated Webpack code is dropped (double that if the custom-panel entry also gets loaded)
- If navigating from the login page, it's already in memory



> Maybe the caching in the server worker has changed, but I was under the impression that it would throw away all the cache of the old service worker version and download a fresh copy.

I assume you're referring to the [`install` event handler](https://github.com/home-assistant/frontend/blob/f4fa0ceef520eab757b3e5f13df84f1b56f209f7/src/entrypoints/service_worker.ts#L181)?  That only deletes the "runtime cache", which does not include everything in the precache manifest.