@steverep 2023-04-12
Perhaps this will help.  Here's a screenshot of the initiator chain for core from a network recording in Chrome dev tools.  Note it's only dependent request is to the web socket API:

![screenshot  of initiator chain for core.js chunk](https://user-images.githubusercontent.com/6835241/231349095-cdce6f63-3eb3-41de-a9c5-e094f8c97752.png)

On the other hand, the app entry needs the runtime to request all the async chunks:

![screenshot  of initiator chain for runtime.js chunk](https://user-images.githubusercontent.com/6835241/231359484-5af76518-cec4-492b-8485-5e793256d4b6.png)

_Please forgive my total lack of precision snipping screenshots!_ 👨🏻‍🦯