@bramkragten 2023-04-12
Splitting the runtime in a single chunk also means we are loading the runtime code for other entrypoints (onboarding, auth, custom panel) while we dont need them in the app? This would increase the code we are loading? Meaning the runtime code of just the core entrypoint is way smaller than the runtime code for all the other entrypoints combined that we now need?