@steverep 2023-04-12
> Splitting the runtime in a single chunk also means we are loading the runtime code for other entrypoints (onboarding, auth, custom panel) while we dont need them in the app? This would increase the code (manifests) we are loading?

Yes, that's true for the manifest part of the runtime, but offset by not redefining things like `function __webpack_require_` and other code multiple times for each entry.  There's also a fair amount of repeats in the manifest since it shares chunks with authorize and onboarding.  Custom panel doesn't load any other chunks.

> Meaning the runtime code of just the core entrypoint is way smaller than the runtime code for all the other entrypoints combined that we now need?

The vast majority of it is already needed because app loads alongside core, so I don't think it's fair to look at the size of its runtime only.

And now to confuse myself... Here are the sizes in KB I get from back to back builds:

| Entry | Current | New |
|:---- |:----:|:----:|
| runtime | n/a | 16 |
| app | 311 | 297 |
| core | 19 | 21 |
| custom-panel | 35 | 39 |
| service_worker | 67 | 67 |
| authorize | 382 | 383 |
| onboarding | 413 | 406 |



The extra couple KB being picked up by some of the entries is baffling me so I'll investigate that.