@balloob 2023-04-13
The page now needs to parse almost twice as much JavaScript (16KB runtime + core is 2KB bigger) before it can do its pre-fetch:

- access local storage to check for existing tokens
- refresh the tokens
- open a websocket connection
- start pre-fetching data needed for the UI. 

And given that most use cases are to show the dashboard etc, current core + app is 330KB, with runtime split out it's 334KB.

I would assume that browsers cache the runtime JS partially parsed. This can make loading that part faster. However, it will always be more work than the current situation, where core.ts doesn't rely on it.