@steverep 2023-04-13
@balloob you're certainly correct - there's more work with the single runtime just for core to arrive at opening the web socket.  Given the small sizes and async nature, I think real world timing differences would be immeasureable, but still, no need to make it worse. 👍🏻 

The size increases in the table above were really bugging me though, so I ran another build with webpack set to give a separate runtime chunk for each entry:

| Entry | Main JS | Runtime |
|:---- |:----:|:----:|
| app | 296K | 16K |
| core | 17K | 2.6K |
| custom-panel | 33K | 2.7K |
| service_worker | 68K | 995 |
| authorize | 376K | 6.8K |
| onboarding | 403K | 11K |

Now those numbers make good sense as they just each basically add up to the current entry sizes.

After digging in a little more, I'm realizing my mistake was not properly factoring in tree shaking and minimization.  When an entry has only its one initial chunk, as is the case for core, custom-panel, and service_worker, Webpack and Terser are able to drop runtime code and compress them significantly more.  Not to mention, if there's only the initial chunk, then there's no caching benefit to separating the runtime anyway.  It'd be nice if Webpack realized that automatically or at least gave a warning.

So I think the solution here to get the caching benefits is to separate the runtime for app, authorize, and onboarding, and leave the others alone.  I'll need to refactor a bit, but that's pretty easy.